requirements.txt
django==1.8.5
dj-database-url==0.3
path.py==7.2
psycopg2==2.5